Refactor Docker workflows and enhance frontend deployment process (#18)
* Update docker.yml

* Update Docker workflow for image publishing

* Change Docker build context to './frontend'

* Refactor Docker workflow for clarity

* Update Docker workflow for frontend image publishing

* Add Docker Compose configuration for frontend service

* Refactor GitHub Actions for Azure deployment

Updated workflow to build, push, and deploy Docker image to Azure App Service.

* Change Docker build context to './frontend'

* Update Azure workflow for container deployment

* Enhance Azure login step with client and tenant IDs

Added additional Azure login parameters for deployment.

* Add id-token permission for OIDC login

* Add auth-type parameter for Azure deployment

* Update Azure login action and credentials method

* Update Azure workflow to lowercase image name

* Update dependencies, remove unused Tailwind CSS packages, and add Docker workflows (#8)

* chore: update dependencies and remove unused Tailwind CSS packages

- Bump React and React DOM versions to 19.2.5
- Remove @tailwindcss/vite and tailwindcss from dependencies
- Remove @tailwindcss/postcss from PostCSS configuration
- Add overrides for postcss version

* feat: add Azure Docker build and deployment workflow

* Upgrade Docker image to npm 11

* Add Docker test workflow and upgrade npm in builds

* Run Docker test workflow on testing-dev pushes

* Delete .github/workflows/azureContainer.yml

* Refactor layout and navigation components; remove unused styles and files

- Updated RootLayout to simplify structure and remove Navbar component.
- Deleted not-found.module.css and moved styles to a new location.
- Adjusted NotFound component to reference new styles path.
- Removed NavLinks and Navbar components, replacing them with new implementations.
- Created new public layout and styles for the public section of the app.
- Added new UI components for navigation and layout in the public section.

* refactor: update layout structure and remove unused metadata; streamline head elements

* refactor: update comment for default page styles to specify public route
