+++ compose.yml
@@ -1,0 +1,15 @@
+version: "3.9"
+
+services:
+  frontend:
+    image: ghcr.io/rgss-cs/williams-rgss-website-dev-frontend-frontend:latest
+    container_name: williams-rgss-frontend
+    restart: unless-stopped
+
+    ports:
+      - "3000:3000"
+
+    environment:
+      NODE_ENV: production
+
+    pull_policy: always
